# tests/contract/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ErroSaida": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErroSaida",
        "type": "object"
      },
...
      "ErrorOrValidationResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          }
        ]
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Structured errors emitted by application and cross-cutting middleware.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "error": {
            "anyOf": [
              {
                "const": "rate_limited",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "retry_after": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "VinculoBancaNovo": {
        "additionalProperties": false,
        "properties": {
          "banca_id": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banca Id"
          }
        },
        "required": [
          "banca_id"
        ],
        "title": "VinculoBancaNovo",
        "type": "object"
      },
      "VinculoBancaSaida": {
        "additionalProperties": false,
        "properties": {
          "banca_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banca Id"
          },
          "conta_casa_id": {
            "title": "Conta Casa Id",
            "type": "integer"
          }
        },
        "required": [
          "conta_casa_id",
          "banca_id"
        ],
        "title": "VinculoBancaSaida",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CollectionToken": {
        "description": "Token opaco da extensão; armazenado no servidor somente como HMAC.",
        "in": "header",
        "name": "X-Coleta-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for authenticated betting, cash-flow, review, upload, collection, and dashboard workflows. The checked-in OpenAPI document is the compatibility contract.",
    "title": "Bancaemdia API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/caixa/contas/{conta_casa_id}/banca": {
      "patch": {
        "description": "Associa ou desassocia uma conta da casa a uma banca do mesmo usuário.",
        "operationId": "vincular_banca_api_v1_caixa_contas__conta_casa_id__banca_patch",
        "parameters": [
          {
            "description": "Identificador da conta da casa usada como filtro.",
            "in": "path",
            "name": "conta_casa_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Conta Casa Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "default": {
                  "summary": "Vincular conta à banca",
                  "value": {
                    "banca_id": 12
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/VinculoBancaNovo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoBancaSaida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroSaida"
                }
              }
            },
            "description": "Conta ou banca não encontrada"
          },
          "413": {
            "content": {
              "text/plain": {
                "example": "Content Too Large",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The request body exceeds the server-wide streaming limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Vincular conta à banca"
      }
    },
    "/api/v1/caixa/extrato": {
      "get": {
        "description": "Combina movimentos de caixa e liquidações de apostas numa linha do tempo paginada.",
        "operationId": "consultar_extrato_api_v1_caixa_extrato_get",
        "parameters": [
          {
            "description": "Identificador da conta da casa usada como filtro.",
            "in": "query",
            "name": "conta_casa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conta Casa Id"
            }
          },
          {
            "description": "Limite inicial inclusivo do intervalo, em ISO 8601.",
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "description": "Limite final exclusivo do intervalo, em ISO 8601.",
            "in": "query",
            "name": "ate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ate"
            }
          },
          {
            "description": "Número da página, começando em 1.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 92233720368547758,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Quantidade máxima de itens retornados na página.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtratoSaida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consultar extrato"
      }
    },
...
  }
}